Module 6_Deployment/requirements.txt
streamlit
tensorflow
opencv-python
numpy
matplotlib
Pillow